requirements.txt
fastapi==0.95.2
pydantic==1.10.7
beautifulsoup4==4.12.0
requests==2.31.0
google-generativeai==0.1.2
openai==0.27.8
